Move KafkaSource E2E tests to helper functions
Allow using KafkaSource E2E tests in the implementations based
on the Kafka Broker.
